refactor(service): streamline core business logic in auth, inventory, cart and order modules; add admin audit for seller listings

+++ src/main/java/com/example/ecommerce_springboot/ecommerce/services/PaymentService.java
@@ -1,11 +1,13 @@
 package com.example.ecommerce_springboot.ecommerce.services;
 
 
-import com.example.ecommerce_springboot.ecommerce.models.User;
+
 import com.razorpay.RazorpayException;
 import com.stripe.exception.StripeException;
 
+import java.math.BigDecimal;
+
 public interface PaymentService {
 
-    String generatePaymentLink(Long orderId, Long amount) throws StripeException, RazorpayException;
+    String generatePaymentLink(Long orderId, BigDecimal amount) throws StripeException, RazorpayException;
 }
